Change "Lab" quantity to "Test" quantity to be more general.
